Fix startup bugs and added more tests (#22)
* Fixed startup by calling websocket on control panel init


* Fixed autosubscribe bug and file change notification registration


* cleanup imports on tests


* Increment clientcore version


* Added tests for notifications, fixed bugs associated with notifications


* Updated file create schema file name


* Updated plugin to use new patch format


* Updated to use patch consolidation


* Updated cliencore, removed int casting, fixed a few bugs


* Updated to use patch manager's notify-on-send


* updated client core version for reliability changes


* Fixed bug where documentListener was not nulled on unfocus


* Auto-save subscribed preferences on each user-initiated change


* Added shadow document


* Re-route property change listener through extension point

* Fix for extension point deadlock issue

* Fixes for editing of closed files


* Added update site


* Added debug message in getFiles


* Changed all modules to 0.0.1 again


* Fix for windows filepath


* 0.0.2


* 0.0.3


* 0.0.4


* Enable logger, 0.0.5


* Fixed bug where missingPatches were not being applied to shadow


* publish 0.0.6


* 0.0.7


* Formatted code, fixed missing-offset bug in DocumentManager, publish 0.0.8


* 0.0.9 - updated logging statement


* 0.0.10 - updated logging statement


* 0.0.11 - Removed a error-throwing debug statement


* Updated appliedDiffs list to use offsetDiff


* 0.0.13 - Updated clientcore for OT fixes


* 0.0.14 - Insertions at the same index should always come before deletions


* 0.0.15 - Patch ignores empty diffs when parsing from string


* 0.0.16 - Disabled shadowdocument updates


* 0.0.17 - Updating clientcore for bug fix in OT case 3d


* 0.0.18 - Restored file shadow updates


* 0.0.19 - send patches in deletion-first order, apply in insertion-first order


* 0.0.20 - Corrected sorting to sort diffs, instead of patches


* 0.0.21 - updated ClientCore for patch tostring fixes


* 0.0.22 - disabled patch sorting, added offset logic for replacements


* 0.0.23 - Updated clientcore to use new OT, patching algorithms, also updated documentManager patching algorithm


* 0.0.24 - Updated ClientCore for no-action patch handling


* 0.0.25 - Update ClientCore for bugfix regarding no-change patches


* 0.0.26 - update 0.0.26


* 0.0.27 - Fixed bug with storing of modified prevDiff


* 0.0.28 - Fixed bugs related to document patching changeList, transformation versions in response handler


* 0.0.29 - corrected bug with string format


* 0.0.30 - Fixed UI bugs


* 0.0.31 - Re-enabled shadow document


* Added category for update site


* Publish 0.0.32 - Configuration strings moved to core


* Moved all versions to category


* 0.0.33 - Various UI fixes/improvements, added documentListener handling for editor content refreshes


* reformatted all code & organized imports


* 0.0.34 - Changed backends to use SSL, updating address


* 0.0.35 - Sends all changes when permissions are changed; allows for read->write change to automatically flush changes


* 0.0.36 - log plugin version at startup

+++ org.code.toboggan.modelmgr/src/org/code/toboggan/modelmgr/extensions/project/ModelMgrProjectGetPermissionConstants.java
@@ -4,10 +4,11 @@ import org.code.toboggan.network.request.extensionpoints.project.IProjectGetPerm
 
 import com.google.common.collect.BiMap;
 
-public class ModelMgrProjectGetPermissionConstants extends AbstractProjectModelMgrHandler implements IProjectGetPermissionConstantsResponse {
+public class ModelMgrProjectGetPermissionConstants extends AbstractProjectModelMgrHandler
+		implements IProjectGetPermissionConstantsResponse {
 
 	@Override
-	public void getPermissionConstants(BiMap<String, Byte> permConstants) {
+	public void getPermissionConstants(BiMap<String, Integer> permConstants) {
 		ss.setPermissionConstants(permConstants);
 	}
 
@@ -15,5 +16,5 @@ public class ModelMgrProjectGetPermissionConstants extends AbstractProjectModelM
 	public void getPermissionConstantsFailed() {
 		// Do nothing
 	}
-	
+
 }
