+++ org.code.toboggan.network/src/org/code/toboggan/network/request/extensionpoints/project/IProjectGetPermissionConstantsResponse.java
@@ -5,6 +5,7 @@ import org.code.toboggan.core.extensionpoints.ICoreExtension;
 import com.google.common.collect.BiMap;
 
 public interface IProjectGetPermissionConstantsResponse extends ICoreExtension {
-	public void getPermissionConstants(BiMap<String, Byte> permConstants);
+	public void getPermissionConstants(BiMap<String, Integer> permConstants);
+
 	public void getPermissionConstantsFailed();
 }
